Implemented missing object description strings for recently added remote objects

+++ Source/ProcessingEngine/ObjectDataHandling/Mux_nA_to_mB/Mux_nA_to_mB.cpp
@@ -107,9 +107,9 @@ bool Mux_nA_to_mB::OnReceivedMessageFromProtocol(ProtocolId PId, RemoteObjectIde
 			jassert(msgData.addrVal.first <= m_protoChCntA);
 			int absChNr = m_protocolAIds.indexOf(PId) * m_protoChCntA + msgData.addrVal.first;
 			int protocolBIndex = absChNr / (m_protoChCntB + 1);
-			int16 chForB = int16(absChNr % m_protoChCntB);
+			int16 chForB = static_cast<int16>(absChNr % m_protoChCntB);
 			if (chForB == 0)
-				chForB = m_protoChCntB;
+				chForB = static_cast<int16>(m_protoChCntB);
 
 			msgData.addrVal.first = chForB;
 			if (m_protocolBIds.size() >= protocolBIndex + 1)
@@ -120,9 +120,9 @@ bool Mux_nA_to_mB::OnReceivedMessageFromProtocol(ProtocolId PId, RemoteObjectIde
 			jassert(msgData.addrVal.first <= m_protoChCntB);
 			int absChNr = m_protocolBIds.indexOf(PId) * m_protoChCntB + msgData.addrVal.first;
 			int protocolAIndex = absChNr / (m_protoChCntA + 1);
-			int16 chForA = int16(absChNr % m_protoChCntA);
+			int16 chForA = static_cast<int16>(absChNr % m_protoChCntA);
 			if (chForA == 0)
-				chForA = m_protoChCntA;
+				chForA = static_cast<int16>(m_protoChCntA);
 
 			msgData.addrVal.first = chForA;
 			if (m_protocolAIds.size() >= protocolAIndex + 1)
